requirements.txt
# 主要依赖
streamlit==1.44.1  # Web界面框架
requests==2.32.3   # HTTP请求库
python-dotenv==1.0.0  # 环境变量管理
tqdm==4.66.1  # 进度条

# 辅助依赖
numpy==1.26.4  # 数值计算库
pandas==2.2.2  # 数据处理
watchdog==4.0.0  # 文件监控（Streamlit热重载需要）
pydantic==2.6.4  # 数据验证

# 文档处理相关
pdf2image==1.16.3  # PDF转图片（可选）
Pillow==10.2.0  # 图像处理
python-docx==1.0.1  # Word文档处理
PyPDF2==3.0.1  # PDF文档处理（可选，作为pdf2image的备选）